--- src/data/service/CourseService.ts
@@ -1,22 +1,0 @@
-import axios from "axios"
-import { ICourse } from "../domain/Course"
-import { ServiceInterface } from "./ServiceInterface"
-import { REST_API_URL } from "../constants"
-import { Response } from "../domain/Response"
-export class CourseService implements ServiceInterface {
-    async getAll(query?: string): Promise<ICourse[]> {
-        const url = query ? `${REST_API_URL}/courses?query=${query}` : `${REST_API_URL}/courses`
-        const courseDTOs = await axios.get<Response<ICourse[]>>(url)
-        console.log(courseDTOs)
-        const courses = courseDTOs.data.data
-        return courses
-    }
-
-    // async getById(id: number): Promise<Course> {
-    //     const courseDTO = await axios.get<Course>(`${REST_API_URL}/courses/${id}`)
-    //     const course = courseDTO.data
-    //     return course
-    // }
-}
-
-export const courseService = new CourseService()
